Fix possible loss of data warnings in RGBController_RoccatKoneAimo.cpp

+++ Controllers/RoccatController/RoccatKoneAimoController/RGBController_RoccatKoneAimo.cpp
@@ -59,88 +59,88 @@ void RGBController_RoccatKoneAimo::SetupZones()
     | Set up zones and leds per zone                            |
     \*---------------------------------------------------------*/
     zone WHEEL_zone;
-    WHEEL_zone.name          = "Scroll Wheel";
-    WHEEL_zone.type          = ZONE_TYPE_SINGLE;
-    WHEEL_zone.leds_min      = 1;
-    WHEEL_zone.leds_max      = 1;
-    WHEEL_zone.leds_count    = 1;
-    WHEEL_zone.matrix_map    = NULL;
+    WHEEL_zone.name                 = "Scroll Wheel";
+    WHEEL_zone.type                 = ZONE_TYPE_SINGLE;
+    WHEEL_zone.leds_min             = 1;
+    WHEEL_zone.leds_max             = 1;
+    WHEEL_zone.leds_count           = 1;
+    WHEEL_zone.matrix_map           = NULL;
     zones.push_back(WHEEL_zone);
     zones_channel.push_back(SCROLL_WHEEL);
 
     led WHEEL_led;
-    WHEEL_led.name = "Wheel LED";
-    WHEEL_led.value = zones.size();
+    WHEEL_led.name                  = "Wheel LED";
+    WHEEL_led.value                 = (unsigned int)zones.size();
     leds.push_back(WHEEL_led);
     leds_channel.push_back(SCROLL_WHEEL);
 
     zone STRIP_LEFT_zone;
-    STRIP_LEFT_zone.name          = "Strip left";
-    STRIP_LEFT_zone.type          = ZONE_TYPE_LINEAR;
-    STRIP_LEFT_zone.leds_min      = 4;
-    STRIP_LEFT_zone.leds_max      = 4;
-    STRIP_LEFT_zone.leds_count    = 4;
-    STRIP_LEFT_zone.matrix_map    = NULL;
+    STRIP_LEFT_zone.name            = "Strip left";
+    STRIP_LEFT_zone.type            = ZONE_TYPE_LINEAR;
+    STRIP_LEFT_zone.leds_min        = 4;
+    STRIP_LEFT_zone.leds_max        = 4;
+    STRIP_LEFT_zone.leds_count      = 4;
+    STRIP_LEFT_zone.matrix_map      = NULL;
     zones.push_back(STRIP_LEFT_zone);
     zones_channel.push_back(STRIP_LEFT);
 
     for(std::size_t led_idx = 0; led_idx < STRIP_LEFT_zone.leds_max; led_idx++)
     {
         led STRIP_LEFT_led;
-        STRIP_LEFT_led.name = "Strip left LED "  + std::to_string(led_idx + 1);
-        STRIP_LEFT_led.value = zones.size();
+        STRIP_LEFT_led.name         = "Strip left LED "  + std::to_string(led_idx + 1);
+        STRIP_LEFT_led.value        = (unsigned int)zones.size();
         leds.push_back(STRIP_LEFT_led);
         leds_channel.push_back(STRIP_LEFT);
     }
 
     zone STRIP_RIGHT_zone;
-    STRIP_RIGHT_zone.name          = "Strip right";
-    STRIP_RIGHT_zone.type          = ZONE_TYPE_LINEAR;
-    STRIP_RIGHT_zone.leds_min      = 4;
-    STRIP_RIGHT_zone.leds_max      = 4;
-    STRIP_RIGHT_zone.leds_count    = 4;
-    STRIP_RIGHT_zone.matrix_map    = NULL;
+    STRIP_RIGHT_zone.name           = "Strip right";
+    STRIP_RIGHT_zone.type           = ZONE_TYPE_LINEAR;
+    STRIP_RIGHT_zone.leds_min       = 4;
+    STRIP_RIGHT_zone.leds_max       = 4;
+    STRIP_RIGHT_zone.leds_count     = 4;
+    STRIP_RIGHT_zone.matrix_map     = NULL;
     zones.push_back(STRIP_RIGHT_zone);
     zones_channel.push_back(STRIP_RIGHT);
 
     for(std::size_t led_idx = 0; led_idx < STRIP_RIGHT_zone.leds_max; led_idx++)
     {
         led STRIP_RIGHT_led;
-        STRIP_RIGHT_led.name = "Strip right LED " + std::to_string(led_idx + 1);
-        STRIP_RIGHT_led.value = zones.size();
+        STRIP_RIGHT_led.name        = "Strip right LED " + std::to_string(led_idx + 1);
+        STRIP_RIGHT_led.value       = (unsigned int)zones.size();
         leds.push_back(STRIP_RIGHT_led);
         leds_channel.push_back(STRIP_RIGHT);
     }
 
     zone LOWER_LEFT_zone;
-    LOWER_LEFT_zone.name          = "Lower left";
-    LOWER_LEFT_zone.type          = ZONE_TYPE_SINGLE;
-    LOWER_LEFT_zone.leds_min      = 1;
-    LOWER_LEFT_zone.leds_max      = 1;
-    LOWER_LEFT_zone.leds_count    = 1;
-    LOWER_LEFT_zone.matrix_map    = NULL;
+    LOWER_LEFT_zone.name            = "Lower left";
+    LOWER_LEFT_zone.type            = ZONE_TYPE_SINGLE;
+    LOWER_LEFT_zone.leds_min        = 1;
+    LOWER_LEFT_zone.leds_max        = 1;
+    LOWER_LEFT_zone.leds_count      = 1;
+    LOWER_LEFT_zone.matrix_map      = NULL;
     zones.push_back(LOWER_LEFT_zone);
     zones_channel.push_back(LOWER_LEFT);
 
     led LOWER_LEFT_led;
-    LOWER_LEFT_led.name = "Lower left LED";
-    LOWER_LEFT_led.value = zones.size();
+    LOWER_LEFT_led.name             = "Lower left LED";
+    LOWER_LEFT_led.value            = (unsigned int)zones.size();
     leds.push_back(LOWER_LEFT_led);
     leds_channel.push_back(LOWER_LEFT);
 
     zone LOWER_RIGHT_zone;
-    LOWER_RIGHT_zone.name          = "Lower right";
-    LOWER_RIGHT_zone.type          = ZONE_TYPE_SINGLE;
-    LOWER_RIGHT_zone.leds_min      = 1;
-    LOWER_RIGHT_zone.leds_max      = 1;
-    LOWER_RIGHT_zone.leds_count    = 1;
-    LOWER_RIGHT_zone.matrix_map    = NULL;
+    LOWER_RIGHT_zone.name           = "Lower right";
+    LOWER_RIGHT_zone.type           = ZONE_TYPE_SINGLE;
+    LOWER_RIGHT_zone.leds_min       = 1;
+    LOWER_RIGHT_zone.leds_max       = 1;
+    LOWER_RIGHT_zone.leds_count     = 1;
+    LOWER_RIGHT_zone.matrix_map     = NULL;
     zones.push_back(LOWER_RIGHT_zone);
     zones_channel.push_back(LOWER_RIGHT);
 
     led LOWER_RIGHT_led;
-    LOWER_RIGHT_led.name = "Lower right LED";
-    LOWER_RIGHT_led.value = zones.size();
+    LOWER_RIGHT_led.name            = "Lower right LED";
+    LOWER_RIGHT_led.value           = (unsigned int)zones.size();
     leds.push_back(LOWER_RIGHT_led);
     leds_channel.push_back(LOWER_RIGHT);
 
@@ -151,11 +151,11 @@ void RGBController_RoccatKoneAimo::SetupZones()
     \*---------------------------------------------------------*/
     for(std::size_t led_idx = 0; led_idx < leds.size(); led_idx++)
     {
-        unsigned char red = 0x00;
-        unsigned char grn = 0x00;
-        unsigned char blu = 0x00;
+        unsigned char red           = 0x00;
+        unsigned char grn           = 0x00;
+        unsigned char blu           = 0x00;
 
-        colors[led_idx] = ToRGBColor(red, grn, blu);
+        colors[led_idx]             = ToRGBColor(red, grn, blu);
     }
 }
